Modify repo architecture
