+++ Makefile
@@ -17,7 +17,7 @@ else
     VECHO = @printf
 endif
 
-OBJS := kvm-host.o
+OBJS := kvm-host.o vm.o
 OBJS := $(addprefix $(OUT)/,$(OBJS))
 deps := $(OBJS:%.o=%.o.d)
 
@@ -25,7 +25,7 @@ $(BIN): $(OBJS)
 	$(VECHO) "  LD\t$@\n"
 	$(Q)$(CC) $(LDFLAGS) -o $@ $^
 
-$(OUT)/%.o: %.c
+$(OUT)/%.o: src/%.c
 	@mkdir -p $(OUT)
 	$(VECHO) "  CC\t$@\n"
 	$(Q)$(CC) -o $@ $(CFLAGS) -c -MMD -MF $@.d $<
