+++ src/vm.h
@@ -1,0 +1,17 @@
+#ifndef VM_H
+#define VM_H
+
+#define RAM_SIZE (1 << 30)
+#define KERNEL_OPTS "console=ttyS0"
+
+typedef struct {
+    int kvm_fd, vm_fd, vcpu_fd;
+    void *mem;
+} vm_t;
+
+int vm_init(vm_t *v);
+int vm_load(vm_t *v, const char *image_path);
+int vm_run(vm_t *v);
+void vm_exit(vm_t *v);
+
+#endif
